abstract ShowingRegistration button element into RegisterButton component
