+++ src/SingleShowing/ShowingRegistration.tsx
@@ -1,22 +1,12 @@
-import { useContext, useState } from "react"
-import { handleRegistration } from "./event-handlers"
-import FirebaseContext from "../../hooks/FirebaseContext"
-import UserContext from "../../hooks/UserContext"
+import { useState } from "react"
+import RegisterButton from "./RegisterButton"
 
 function ShowingRegistration({id}: {id: string}) {
-    const [isButtonDisabled, setIsButtonDisabled] = useState(false)
-    const firebase = useContext(FirebaseContext)
-    const {user: {username}} = useContext(UserContext)
-    const [isError, setIsError] = useState(false) 
+    const [isError, setIsError] = useState(true)
 
     return <>
         <p>Interested in attending?</p>
-        <button 
-            onClick={() => handleRegistration(setIsButtonDisabled, setIsError, firebase, username, id)} 
-            disabled={isButtonDisabled}
-        >
-            Register
-        </button>
+        <RegisterButton setIsError={setIsError} id={id} />
         {isError ? <p>Error occurred during registration, please try again</p> : null}
     </>
 }
